refactor(issues): create issues store, move issues logic to it
